Revert "Everything goes in a module. Major refactor."
