--- discordbot/README.md
@@ -1,37 +1,0 @@
-
-# Penny-Dreadful-Discord-Bot
-Displays info about quoted cards in a discord channel
-
-[![Build Status](http://ci.katelyngigante.com/buildStatus/icon?job=Penny Dreadful/Penny-Dreadful-Discord-Bot/master)](http://ci.katelyngigante.com/job/Penny%20Dreadful/job/Penny-Dreadful-Discord-Bot/job/master/)
-
-#Usage
-Basic bot usage: Include [cardname] in your regular messages.
-The bot will search for any quoted cards, and respond with the card details.
-
-#Commands
-
-`!oracle {name}` Give the Oracle text of the named card.
-`!barbs` Gives Volvary's helpful advice for when to sideboard in Aura Barbs.
-`!help` Provides information on how to operate the bot.
-`!modofail` Ding!
-`!price {name}` Get price information about the named card.
-`!quality` A helpful reminder about everyone's favorite way to play digital Magic
-`!random` Request a random PD legal card
-`!random X` Request X random PD legal cards.
-`!resources` Link to page of all Penny Dreadful resources.
-           `!resources {section}` Link to Penny Dreadful resources section.
-           `!resources {section} {link}` Link to Penny Dreadful resource.
-        
-`!rhinos` Anything can be a rhino if you try hard enough
-`!rotation` Give the date of the next Penny Dreadful rotation.
-`!search {query}` Search for cards, using a magidex style query.
-`!status` Gives the status of MTGO, UP or DOWN.
-
-#Installation
-To add this bot to your servers use this <a href='https://discordapp.com/oauth2/authorize?client_id=224755717767299072&scope=bot&permissions=0'>link</a>
-
-#Libraries used
-
-[Discord.py](https://github.com/Rapptz/discord.py)
-
-[mtgjson](http://mtgjson.com/)
