+++ emoji.py
